backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
openpyxl==3.1.5
psycopg2-binary==2.9.9
pgvector==0.3.5
openai==1.59.3  # OpenAI SDK pointed at OpenRouter — 1.52+ required for httpx 0.28 compatibility
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
json-repair==0.30.3  # fixes malformed JSON from LLMs (unescaped quotes in reasoning fields)
